fixed a bug in multiplayer where requestUpdate() was only working for the first client

+++ OgrianOrientedSprite.cpp
@@ -160,7 +160,7 @@ int OrientedSprite::addPose(String basename)
 
 void OrientedSprite::setPose(int index)
 {
-	if (index >= mPoses.size())
+	if (index >= (int)mPoses.size())
 		Except( Exception::ERR_INTERNAL_ERROR, "Error: Pose not found.",
 				"OrientedSprite::setPose" );
